Use better GC settings
Lower the value of gc-cons-threshold, so that the GC runs reasonably often when
the heap is small to minimize the GC pause time, but increase gc-cons-percentage
so that big heaps are scanned less often. This should result in lower
interactive latency *and* higher throughput that the old settings.

This would be unnecessary had Emacs had a real generational GC; alas, it's still
the same, naive STW M&S GC.